+++ src/charm.py
@@ -28,7 +28,6 @@ from state import (
     State,
 )
 
-AGENT_DISCOVERY_INGRESS_RELATION_NAME = "agent-discovery-ingress"
 INGRESS_RELATION_NAME = "ingress"
 logger = logging.getLogger(__name__)
 
@@ -56,13 +55,17 @@ class JenkinsK8sOperatorCharm(ops.CharmBase):
 
         # Ingress dedicated to agent discovery
         self.agent_discovery_ingress_observer = ingress.Observer(
-            self, "agent-discovery-ingress-observer", AGENT_DISCOVERY_INGRESS_RELATION_NAME
+            self, "agent-discovery-ingress-observer", agent.AGENT_DISCOVERY_INGRESS_RELATION_NAME
         )
         self.ingress_observer = ingress.Observer(self, "ingress-observer", INGRESS_RELATION_NAME)
         self.jenkins = jenkins.Jenkins(self.calculate_env())
         self.actions_observer = actions.Observer(self, self.state, self.jenkins)
         self.agent_observer = agent.Observer(
-            self, self.state, self.agent_discovery_ingress_observer, self.jenkins
+            self,
+            self.state,
+            self.agent_discovery_ingress_observer,
+            self.ingress_observer,
+            self.jenkins,
         )
         self.cos_observer = cos.Observer(self)
         self.auth_proxy_observer = auth_proxy.Observer(