+++ docs/changelog.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 Each revision is versioned by the date of the revision.
 
+### 2025-09-05
+
+- Fix issue with Jenkins agent node server discovery when ingress is applied to server
+
 ### 2025-09-02
 
 - Upgrade Jenkins version to latest LTS (v2.516.2)